Don't show unused references in "make cite"
This invokes checkcites with the "--undefined" flag, which means that
only undefined citations in the TeX sources will be shown, not unused
references in the bibliography file. This is useful for bibliography
files which are managed by third party applications such as Mendeley, in
with which a single bibliography file can track many hundreds of
academic papers.