proposal: Keep a local copy of needed references
Rather than relying entirely on a large Mendeley-managed bibtex file
which isn't tracket in this repository, pull out the required citations
at compile time and store them in a locally tracked refs.bib file, to
support building with or without the library bibfile. If the global
bibfile isn't present, use the local version.