Use pdflatex to generate list of input sources
This adds the "-recorder" argument to the invocation of pdflatex so that
it generates a .fls which records (amongst other things) *all* input
files which are read. We can use this to derive an accurate list of
dependencies without relying on each document's Makefile to provide a
list of sources.

If no .fls file exists, we simply assume that the target .tex file is
the only dependency.