Makefile.latex: (wc|cite|lint) depend on PDF
This makes $(DocumentPDF) a prerequisite for generating word counts,
checking citation accuracies, or running the lint checks, since we
require up to date .fls or .aux files for each. These .fls and .aux
files are generated by compiling the PDF.

The reason for the shell conditional instead of simply adding
$(DocumentPDF) to the list of prerequisites is that we must defer
expansion of the $(DocumentSources) variable until $(DocumentPDF) has
been compiled, so that we can ensure that the .fls and .aux files have
been generated.