configure: Bump required automake version
The AM_EXTRA_RECURSIVE_TARGETS macro used in configure.ac is a new
feature of automake < 1.13. This patch bumps the required automake
version to address this. This is to fix a configure error when building
with Ubuntu, where the distro version of automake is 1.11.3.

See GNU Automake 1.13 release notes:

    http://lwn.net/Articles/531373/

Issue #13.