configure: Allow user-specified compilers
This adds a --disable-local-cc flag to the configure script which allows
users to specify their own compiler toolchains instead of the local copy
of llvm. Users may pass the CC and CXX arguments to the configure
script.

This requires extra conditional logic so that compilers won't be invoked
with unsupported features, such as -stdlib=<> for GCC.

Issue #24.