Add a hardware description file
Currently it contains just the output of `mkopenclinfo`, but this will
be extended to include a description of the host hardware. Future plans
are to implement a `get_hw_features()` function which will parse the
output of a SkelCL program and determine a description of the underlying
hardware, to replace the current "hostname", "dev_type", and "dev_count"
features in e14.